+++ api/src/database.ts
@@ -5,6 +5,7 @@ import { Username } from '@/models/username.model';
 import { ChatMessage } from '@/models/message.model';
 import { TwitterPost } from '@/models/twitter-posts.model';
 import { Repo } from '@/models/repos.model';
+import { JobState } from '@/models/job.model';
 
 export const AppDataSource = new DataSource({
   type: 'mysql',
@@ -13,7 +14,7 @@ export const AppDataSource = new DataSource({
   username: process.env.MYSQLUSER || 'root',
   password: process.env.DB_PASSWORD || 'password',
   database: process.env.MYSQL_DATABASE || 'repo-ranger',
-  entities: [Username, ChatMessage, TwitterPost, Repo, Event],
+  entities: [Username, ChatMessage, TwitterPost, Repo, Event, JobState],
   synchronize: true,
   logging: false,
 });